+++ Changelog.md
@@ -34,5 +34,6 @@
 0.3.4 performance test
 0.3.5 transaction
 0.3.6 read uncommited transaction level
-0.3.7 repeatable read and seriazable transaction level
+0.3.7 repeatable read  transaction level
 0.3.8 config
+0.3.9 SERIALIZABLE transaction level